requirements.txt
numpy==1.16.4
Pillow==5.0.0